+++ python/python/glide/async_commands/core.py
@@ -1491,8 +1491,8 @@ class CoreCommands(Protocol):
         Examples:
             >>> await client.zrange("my_sorted_set", RangeByIndex(0, -1))
                 ['member1', 'member2', 'member3']  # Returns all members in ascending order.
-            >>> await client.zrange("my_sorted_set", RangeByScore(start=InfBound.NEG_INF, stop= ScoreBoundary(3), limit= Limit(0 , 2)))
-                ['member2', 'member3'] # Returns members with scores within the range of negative infinity to 3, limited to 2 elements, in ascending order.
+            >>> await client.zrange("my_sorted_set", RangeByScore(start=InfBound.NEG_INF, stop=ScoreBoundary(3)))
+                ['member2', 'member3'] # Returns members with scores within the range of negative infinity to 3, in ascending order.
         """
         args = _create_zrange_args(key, range_query, reverse, with_scores=False)
 
@@ -1524,8 +1524,8 @@ class CoreCommands(Protocol):
         Examples:
             >>> await client.zrange_withscores("my_sorted_set", RangeByScore(ScoreBoundary(10), ScoreBoundary(20)))
                 {'member1': 10.5, 'member2': 15.2}  # Returns members with scores between 10 and 20 with their scores.
-            >>> await client.zrange("my_sorted_set", RangeByScore(start=InfBound.NEG_INF, stop= ScoreBoundary(3), limit= Limit(0 , 2)))
-                {'member4': -2.0, 'member7': 1.5} # Returns members with with scores within the range of negative infinity to 3, limited to 2 elements, with their scores.
+           >>> await client.zrange("my_sorted_set", RangeByScore(start=InfBound.NEG_INF, stop=ScoreBoundary(3)))
+                {'member4': -2.0, 'member7': 1.5} # Returns members with with scores within the range of negative infinity to 3, with their scores.
         """
         args = _create_zrange_args(key, range_query, reverse, with_scores=True)